+++ src/app.h
@@ -1,0 +1,5 @@
+#pragma once
+
+ namespace app {
+    int run();
+}
