+++ src/systems.h
@@ -1,0 +1,4 @@
+#include "input.h"
+#include "audio.h"
+#include "events.h"
+#include "asset_loader.h"
