+++ src/ui.cpp
@@ -7,6 +7,11 @@
 
 #include "debug_macros.h"
 
+const char* StringPrison::add(std::string&& string) {
+    strings.push_back(std::move(string));
+
+    return strings.back().data();
+}
 
 UI::UI(int _screen_width, int _screen_height) : screen_width{ _screen_width }, screen_height{ _screen_height } {}
